Refactor: #175 닉네임 중복체크 훅 분리
